add serverInterface
