+++ T1/README.md
@@ -1,6 +1,6 @@
 How to compile
 ```
-g++ -Iinclude -o servidor server/main.cpp server/Server.cpp client/Client.cpp
+g++ -Iinclude -o servidor server/main.cpp server/Server.cpp server/serverInterface.cpp client/Client.cpp
 ```
 
 How to init server
